Add batch mode to reuse loaded keystore
Accepts commands as stdin arguments separated by NUL (\0), preceded
by an argument count in string form. Outputs NUL when a given command
has been processed, followed by a return value in string form, where 0
is success. This is sent to both stdout and stderr to mark the
completion of a command.

This change is intended to enhance performance by keeping the keystore
loaded, as PKCS#11 keystores may take a significant time to load
depending on the implementation.

Change-Id: Ic140cf96509707a2a8b720727fdf1dc5e636e3db